src/main/java/kz/iitu/itse1909r_spring/ushurbakiev_davlatbek_itse1909r/Controller/SwaggerController.java
package kz.iitu.itse1909r_spring.ushurbakiev_davlatbek_itse1909r.Controller;

import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class SwaggerController {
    //http://localhost:8080/swagger-ui.html#/
    @RequestMapping(method = RequestMethod.GET, value = "/api/javainuse")
    public String sayHello() {
        return "Swagger Hello World";
    }
}
